streamlit/requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
plotly>=5.0.0
httpx>=0.25.0
python-dotenv>=1.0.0